# openapi.yaml
openapi: 3.0.3
info:
  title: Hadi Hep Beraber - Okey Game API
  description: API for Turkish Okey game platform
  version: 1.0.0
servers:
  - url: /api
    description: Next.js API routes
...
paths:
...
  /games/{gameId}/leave:
    post:
      tags: [games]
      summary: Leave a game (before it starts)
      security:
        - session: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Left successfully
...
  /games/{gameId}/draw:
    post:
      tags: [game-actions]
      summary: Draw a tile
      security:
        - session: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source]
              properties:
                source:
                  type: string
                  enum: [pile, discard]
      responses:
        '200':
          description: Tile drawn
          content:
            application/json:
              schema:
                type: object
                properties:
                  tile:
                    $ref: '#/components/schemas/Tile'
        '400':
          description: Not your turn or wrong phase
...
  /games/{gameId}/finish:
    post:
      tags: [game-actions]
      summary: Declare win
      security:
        - session: []
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [discardTileId]
              properties:
                discardTileId:
                  type: string
                  description: Tile to discard when finishing
      responses:
        '200':
          description: Win declared successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  winner:
                    type: string
                  scores:
                    type: array
                    items:
                      type: object
                      properties:
                        playerId:
                          type: string
                        score:
                          type: integer
        '400':
          description: Invalid winning hand
...
  /games/{gameId}/replay:
    get:
      tags: [game-events]
      summary: Get replay data
      description: |
        Returns complete replay data including initial state and all events.
        Only available for finished games.
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  game:
                    $ref: '#/components/schemas/Game'
                  initialState:
                    type: object
                    description: Game state at start (hands, tileBag, etc.)
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/GameEvent'
        '400':
          description: Game not finished yet

  # ============================================
  # FRIENDS
  # ============================================
...
  /friends/{userId}/invite:
    post:
      tags: [friends]
      summary: Invite friend to your game
      security:
        - session: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [gameId]
              properties:
                gameId:
                  type: string
      responses:
        '200':
          description: Invite sent

  # ============================================
  # SHOP
  # ============================================
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
...
    Tile:
      type: object
      properties:
        id:
          type: string
        number:
          type: integer
          minimum: 1
          maximum: 13
        color:
          type: string
          enum: [red, yellow, blue, black]
        tileType:
          type: string
          enum: [normal, fake_okey]
          description: |
            - normal: Regular numbered tile
            - fake_okey: Sahte okey (joker tile) - represents the real okey's value only, NOT a wildcard
            Note: Whether a tile IS the real okey is determined dynamically (indicator + 1)
...
    GameRules:
      type: object
      description: Rule configuration for a game variant
      properties:
        variant:
          type: string
          enum: [okey, okey101]
          default: okey
        openingRequired:
          type: boolean
          default: false
          description: Must open (show valid groups) before finishing
        openingMinScore:
          type: integer
          default: 0
          description: Minimum score to open (101 variant)
        allowCift:
          type: boolean
          default: true
          description: Allow çift (7 pairs) win
        timeoutPolicy:
          type: string
          enum: [auto_discard_random, auto_discard_last_drawn, skip_turn]
          default: auto_discard_random
          description: What happens when turn times out
    GameEvent:
      type: object
      description: Immutable game event for replay and audit
      properties:
        id:
          type: string
        gameId:
          type: string
        type:
          type: string
          enum:
            - GAME_STARTED
            - DRAW_PILE
            - DRAW_DISCARD
            - DISCARD
            - FINISH
            - TURN_TIMEOUT
            - PLAYER_DISCONNECTED
            - PLAYER_RECONNECTED
        actorPlayerId:
          type: string
        payload:
          type: object
          description: Event-specific data (tile drawn, discarded, etc.)
        createdAt:
          type: string
          format: date-time
    Game:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [WAITING, STARTING, PLAYING, FINISHED, CANCELLED]
        mode:
          type: string
          enum: [regular, okey101]
        maxPlayers:
          type: integer
        isPrivate:
          type: boolean
        roomCode:
          type: string
        players:
          type: array
          items:
            $ref: '#/components/schemas/GamePlayer'
        currentTurn:
          type: integer
        turnPhase:
          type: string
          enum: [draw, discard]
        indicatorTile:
          $ref: '#/components/schemas/Tile'
        okeyTile:
          $ref: '#/components/schemas/Tile'
        discardPileTop:
          $ref: '#/components/schemas/Tile'
        tileBagCount:
          type: integer
        turnTimeLimit:
          type: integer
        turnStartedAt:
          type: string
          format: date-time
        rules:
          $ref: '#/components/schemas/GameRules'
        stake:
          type: object
          properties:
            entryFee:
              type: integer
            pot:
              type: integer
            potDistribution:
              type: string
        winnerId:
          type: string
        finishType:
          type: string
          enum: [normal, cift, el]
          description: |
            - normal: Standard finish
            - cift: 7 pairs (2x score in 101)
            - el: First turn win (no draw)
...
    GamePlayer:
      type: object
      properties:
        id:
          type: string
        position:
          type: integer
        name:
          type: string
        avatar:
          type: string
        isAI:
          type: boolean
        isConnected:
          type: boolean
        isReady:
          type: boolean
        tileCount:
          type: integer
          description: Number of tiles (tiles themselves hidden for opponents)
        tiles:
          type: array
          items:
            $ref: '#/components/schemas/Tile'
          description: Only included for the requesting player
...